requirements.txt
setuptools
numpy
torch
torchvision
openai-whisper
pyannote.audio
soundfile
speechbrain==0.5.16

# Server dependencies
fastapi
uvicorn
python-multipart
